# Conflicts:
#	src/main/java/com/mall/wms/comm/CodeMsg.java

+++ src/main/java/com/mall/wms/entity/UserEntity.java
@@ -90,7 +90,7 @@ public class UserEntity implements Serializable {
     /**
      * 注册时间戳
      */
-    private Integer regTime;
+    private Long regTime;
 
     /**
      * 添加时间
